Fix QR code generation on initial request
In case of initial QR code wasn't scanned withing defined TTL new QR
code has to be generated. For that initial request has to be sent again
to get new ref, otherwise the same QR code will be generated over and
over again and user could not link it in the Whastapp mobile app, as it
fails with QR code outdated error.